+++ aeneas/cdtw/cdtw_py.c
@@ -5,7 +5,7 @@
 #
 # Copyright (C) 2012-2013, Alberto Pettarin (www.albertopettarin.it)
 # Copyright (C) 2013-2015, ReadBeyond Srl   (www.readbeyond.it)
-# Copyright (C) 2015-2017, Alberto Pettarin (www.albertopettarin.it)
+# Copyright (C) 2015-2018, Alberto Pettarin (www.albertopettarin.it)
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as published by
@@ -36,15 +36,17 @@ Python C Extension for computing the DTW
 // append a new tuple (i, j) to the given list
 static void _append(PyObject *list, uint32_t i, uint32_t j) {
     PyObject *tuple;
-    
+
     tuple = PyTuple_New(2);
     // PyTuple_SetItem steals a reference, so no PyDECREF is needed
     PyTuple_SetItem(tuple, 0, Py_BuildValue("I", i));
     PyTuple_SetItem(tuple, 1, Py_BuildValue("I", j));
     PyList_Append(list, tuple);
+
+    Py_DECREF(tuple);
 }
 
-// convert array of struct to list of tuples 
+// convert array of struct to list of tuples
 static void _array_to_list(struct PATH_CELL *best_path, uint32_t best_path_length, PyObject *list) {
     uint32_t k;
 
@@ -63,7 +65,7 @@ static PyObject *compute_best_path(PyObject *self, PyObject *args) {
     PyObject *mfcc1_raw;
     PyObject *mfcc2_raw;
     uint32_t delta;
- 
+
     PyArrayObject *mfcc1, *mfcc2, *cost_matrix, *centers;
     PyObject *best_path_ptr;
     npy_intp cost_matrix_dimensions[2];
@@ -103,15 +105,15 @@ static PyObject *compute_best_path(PyObject *self, PyObject *args) {
         return NULL;
     }
 
-    // delta cannot be greater than m 
+    // delta cannot be greater than m
     if (delta > m) {
         delta = m;
     }
-    
+
     // pointer to cost matrix data
     mfcc1_ptr = (double *)PyArray_DATA(mfcc1);
     mfcc2_ptr = (double *)PyArray_DATA(mfcc2);
-    
+
     // create cost matrix object
     cost_matrix_dimensions[0] = n;
     cost_matrix_dimensions[1] = delta;
@@ -132,7 +134,7 @@ static PyObject *compute_best_path(PyObject *self, PyObject *args) {
        PyErr_SetString(PyExc_ValueError, "Error while computing cost matrix");
        return NULL;
     }
-    
+
     if (_compute_accumulated_cost_matrix_in_place(cost_matrix_ptr, centers_ptr, n, delta) != CDTW_SUCCESS) {
        Py_XDECREF(mfcc1);
        Py_XDECREF(mfcc2);
@@ -141,7 +143,7 @@ static PyObject *compute_best_path(PyObject *self, PyObject *args) {
        PyErr_SetString(PyExc_ValueError, "Error while computing accumulated cost matrix");
        return NULL;
     }
-    
+
     if (_compute_best_path(cost_matrix_ptr, centers_ptr, n, delta, &best_path, &best_path_length) != CDTW_SUCCESS) {
        Py_XDECREF(mfcc1);
        Py_XDECREF(mfcc2);
@@ -151,7 +153,7 @@ static PyObject *compute_best_path(PyObject *self, PyObject *args) {
        return NULL;
     }
 
-    // convert array of struct to list of tuples 
+    // convert array of struct to list of tuples
     best_path_ptr = PyList_New(0);
     _array_to_list(best_path, best_path_length, best_path_ptr);
     free((void *)best_path);
@@ -166,7 +168,7 @@ static PyObject *compute_best_path(PyObject *self, PyObject *args) {
     return best_path_ptr;
 }
 
-// compute the cost matrix and the corresponding stripe centers 
+// compute the cost matrix and the corresponding stripe centers
 // take the PyObject containing the following arguments:
 //   - mfcc1:       2D array (l x n) of double, MFCCs of the first wave
 //   - mfcc2:       2D array (l x m) of double, MFCCs of the second wave
@@ -186,7 +188,7 @@ static PyObject *compute_cost_matrix_step(PyObject *self, PyObject *args) {
     double *mfcc1_ptr, *mfcc2_ptr, *cost_matrix_ptr;
     uint32_t *centers_ptr;
     uint32_t l1, l2, n, m;
-   
+
     // O = object (do not convert or check for errors)
     // I = unsigned int
     if (!PyArg_ParseTuple(args, "OOI", &mfcc1_raw, &mfcc2_raw, &delta)) {
@@ -216,7 +218,7 @@ static PyObject *compute_cost_matrix_step(PyObject *self, PyObject *args) {
         return NULL;
     }
 
-    // delta cannot be greater than m 
+    // delta cannot be greater than m
     if (delta > m) {
         delta = m;
     }
@@ -235,7 +237,7 @@ static PyObject *compute_cost_matrix_step(PyObject *self, PyObject *args) {
     centers_dimensions[0] = n;
     centers = (PyArrayObject *)PyArray_SimpleNew(1, centers_dimensions, NPY_UINT32);
     centers_ptr = (uint32_t *)PyArray_DATA(centers);
-    
+
     // compute cost matrix
     if (_compute_cost_matrix(mfcc1_ptr, mfcc2_ptr, delta, cost_matrix_ptr, centers_ptr, n, m, l1) != CDTW_SUCCESS) {
         Py_XDECREF(mfcc1);
@@ -251,7 +253,7 @@ static PyObject *compute_cost_matrix_step(PyObject *self, PyObject *args) {
     Py_DECREF(mfcc2);
 
     // return tuple with computed cost matrix and centers
-    // PyTuple_SetItem steals a reference, so no PyDECREF is needed 
+    // PyTuple_SetItem steals a reference, so no PyDECREF is needed
     tuple = PyTuple_New(2);
     PyTuple_SetItem(tuple, 0, PyArray_Return(cost_matrix));
     PyTuple_SetItem(tuple, 1, PyArray_Return(centers));
@@ -287,7 +289,7 @@ static PyObject *compute_accumulated_cost_matrix_step(PyObject *self, PyObject *
 
     // pointer to cost matrix data
     cost_matrix_ptr = (double *)PyArray_DATA(cost_matrix);
-    
+
     // get the dimensions of the input arguments
     n = PyArray_DIMS(cost_matrix)[0];
     delta = PyArray_DIMS(cost_matrix)[1];
@@ -297,11 +299,11 @@ static PyObject *compute_accumulated_cost_matrix_step(PyObject *self, PyObject *
         PyErr_SetString(PyExc_ValueError, "The number of rows of cost_matrix must be equal to the number of elements of centers");
         return NULL;
     }
-   
+
     // pointer to centers data
     centers_ptr = (uint32_t *)PyArray_DATA(centers);
-    
-    // create accumulated cost matrix object 
+
+    // create accumulated cost matrix object
     accumulated_cost_matrix_dimensions[0] = n;
     accumulated_cost_matrix_dimensions[1] = delta;
     accumulated_cost_matrix = (PyArrayObject *)PyArray_SimpleNew(2, accumulated_cost_matrix_dimensions, NPY_DOUBLE);
@@ -325,7 +327,7 @@ static PyObject *compute_accumulated_cost_matrix_step(PyObject *self, PyObject *
     return PyArray_Return(accumulated_cost_matrix);
 }
 
-// compute the best path 
+// compute the best path
 // take the PyObject containing the following arguments:
 //   - accumulated_cost_matrix: 2D array (n x delta) of double
 //   - centers:                 1D array (n x 1) of int, centers[i] is the 0 <= center < m of the stripe at row i
@@ -354,7 +356,7 @@ static PyObject *compute_best_path_step(PyObject *self, PyObject *args) {
 
     // pointer to cost matrix data
     accumulated_cost_matrix_ptr = (double *)PyArray_DATA(accumulated_cost_matrix);
-    
+
     // get the dimensions of the input arguments
     n = PyArray_DIMS(accumulated_cost_matrix)[0];
     delta = PyArray_DIMS(accumulated_cost_matrix)[1];
@@ -364,13 +366,13 @@ static PyObject *compute_best_path_step(PyObject *self, PyObject *args) {
         PyErr_SetString(PyExc_ValueError, "The number of rows of accumulated_cost_matrix must be equal to the number of elements of centers");
         return NULL;
     }
-   
+
     // pointer to centers data
     centers_ptr = (uint32_t *)PyArray_DATA(centers);
-    
+
     // create best path array of integers
     best_path_ptr = PyList_New(0);
-    
+
     // compute best path
     if (_compute_best_path(accumulated_cost_matrix_ptr, centers_ptr, n, delta, &best_path, &best_path_length) != CDTW_SUCCESS) {
         Py_XDECREF(accumulated_cost_matrix);
@@ -379,7 +381,7 @@ static PyObject *compute_best_path_step(PyObject *self, PyObject *args) {
         return NULL;
     }
 
-    // convert array of struct to list of tuples 
+    // convert array of struct to list of tuples
     _array_to_list(best_path, best_path_length, best_path_ptr);
     free((void *)best_path);
     best_path = NULL;
@@ -472,7 +474,7 @@ static PyObject *moduleinit(void) {
 PyMODINIT_FUNC PyInit_cdtw(void) {
     PyObject *ret = moduleinit();
     import_array();
-    return ret; 
+    return ret;
 }
 #else
 PyMODINIT_FUNC initcdtw(void) {