+++ docs/source/changelog.rst
@@ -5,6 +5,17 @@ v1.7.3.1 (2020-05-06)
 #. Fixed spurious warnings about not using UTF-8 when "utf-8" is seen
 #. Added debian packaging for python3-aeneas
 
+v1.7.4 (2018-05-??)
+-------------------
+
+#. Fixed bug #203 (support for macOS TTS)
+#. Fixed bugs #177 and #191 (installation issues with numpy)
+#. Fixed bug #205 (memory leak in Python C extension)
+#. Minor documentation improvements
+#. Updated copyright strings with 2018
+#. Updated acknowledgements
+#. Added Travis CI (Linux, macOS)
+
 v1.7.3 (2017-03-15)
 -------------------